requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic==2.7.1